chore(system): remove dedicated JdbcServiceLivenessManager
