+++ jdbc/src/test/java/io/kestra/jdbc/server/JdbcServiceLivenessManagerTest.java
@@ -2,15 +2,7 @@ package io.kestra.jdbc.server;
 
 import io.kestra.core.contexts.KestraContext;
 import io.kestra.core.models.ServerType;
-import io.kestra.core.server.ServerConfig;
-import io.kestra.core.server.ServerInstanceFactory;
-import io.kestra.core.server.Service;
-import io.kestra.core.server.ServiceInstance;
-import io.kestra.core.server.LocalServiceStateFactory;
-import io.kestra.core.server.ServiceLivenessUpdater;
-import io.kestra.core.server.ServiceRegistry;
-import io.kestra.core.server.ServiceStateTransition;
-import io.kestra.core.server.WorkerTaskRestartStrategy;
+import io.kestra.core.server.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -37,7 +29,7 @@ class JdbcServiceLivenessManagerTest {
     @Mock
     public ServiceLivenessUpdater serviceLivenessUpdater;
 
-    private JdbcServiceLivenessManager serviceLivenessManager;
+    private ServiceLivenessManager serviceLivenessManager;
 
     @Mock
     private KestraContext context;
@@ -59,7 +51,7 @@ class JdbcServiceLivenessManagerTest {
             )
         );
 
-        this.serviceLivenessManager = new JdbcServiceLivenessManager(
+        this.serviceLivenessManager = new ServiceLivenessManager(
             config,
             new ServiceRegistry(),
             new LocalServiceStateFactory(config, null),
